+++ dcor/_fast_dcov_mergesort.py
@@ -126,7 +126,7 @@ def _compute_row_sums(x):
     return term_1 + term_2
 
 
-def _distance_covariance_mergesort_generic(x, y, unbiased=False):
+def _distance_covariance_sqr_mergesort_generic(x, y, unbiased=False):
 
     n = len(x)
 
@@ -159,3 +159,13 @@ def _distance_covariance_mergesort_generic(x, y, unbiased=False):
              a_dot_dot / n * b_dot_dot / d1 / d2 / d3)
 
     return d_cov
+
+
+def _distance_covariance_sqr_mergesort(x, y):
+    """Fast algorithm for the biased squared distance covariance"""
+    return _distance_covariance_sqr_mergesort_generic(x, y, unbiased=False)
+
+
+def _u_distance_covariance_sqr_mergesort(x, y):
+    """Fast algorithm for the unbiased squared distance covariance"""
+    return _distance_covariance_sqr_mergesort_generic(x, y, unbiased=True)
